ref: move crawler into its own module
